Cleaned up the code removing todos, duplicate code, comments, etc. Finished initial functional config generation for controller, edge router, and edge router using WSS Refactored some existing code to better organize it
